+++ sources/packages/web/src/types/formio.ts
@@ -10,6 +10,7 @@ export interface FormIOForm {
     row: any,
     silent: boolean,
   ) => boolean;
+  submit: () => any;
 }
 
 /**
